refactor(096-W3): FeishuListener 拆 3 类——FeishuApiClient/SlashCommandHandler/CardActionHandler（listener 2600→970 行）
- 设计依据 docs/design/100（拆分边界与字段归属表）；82 函数归域：
  E 发送凭证族 12 → feishu_api_client.rs（unit struct 关联函数，无状态）
  C 斜杠命令族 11 → feishu_slash_commands.rs；D 卡片回调族 24 → feishu_card_actions.rs
  A/B 连接管理+消息编排 21 函数留守（仅它们持有 &self 状态 channels/debounce）
- 纯搬迁零逻辑改动（函数名集合 80=80 守恒、#[test] 7=7、代表函数逐字一致）
- 两处职责修正：send_raw/send_card_raw 从 &self 方法改显式接参关联函数，
  feishu_push 调用点同步改 FeishuApiClient::（顺带缓解 C2 点名的 DIP 违反）；
  FeishuCommandContext/ActionOutcome 字段提 pub(crate)（跨模块流转需要）
- clippy 零告警、全量 1676 测试过、dev 实例启动装配正常（agent_bots 空，
  无真实凭证，WS 通路冒烟以集成测试+装配验证为下限，见 PR 说明）
