+++ backend/src/services/feishu_listener.rs
@@ -2,7 +2,6 @@ use dashmap::DashMap;
 use std::sync::Arc;
 use tokio::sync::mpsc;
 
-use crate::feishu::sdk::config::Config as FeishuSdkConfig;
 use crate::feishu::sdk::token_manager::TokenManager;
 use crate::feishu::{
     create_channel, ChannelMessage, FeishuChannelService, FeishuConfig, FeishuConnectionMode,
@@ -12,12 +11,10 @@ use crate::feishu::{
 use crate::service_context::ServiceContext;
 use crate::task_manager::TaskManager;
 use crate::db::{Database, NewFeishuMessage};
-use crate::services::feishu_card::{
-    Card, CardElement, CardMarkdown, ExecutorOption, HelpCardState, HistoryItem, LoopItem, RecentTaskItem,
-    TodoItem, WorkspaceItem, WorkspaceSummary, build_help_console_card, build_history_card,
-    render_card,
-};
 use crate::models::{AgentBot, BotConfig, build_trigger_params};
+use crate::services::feishu_api_client::FeishuApiClient;
+use crate::services::feishu_card_actions::CardActionHandler;
+use crate::services::feishu_slash_commands::SlashCommandHandler;
 use crate::services::message_debounce::{MessageDebounce, PendingMessage};
 
 /// Manages WebSocket connections to Feishu for all bound bots.
@@ -44,19 +41,19 @@ pub(crate) struct ListenerMessageContext<'a> {
     pub(crate) ctx: &'a ServiceContext,
 }
 
-struct FeishuCommandContext<'a> {
-    db: &'a Arc<Database>,
-    credentials: &'a DashMap<i64, (String, String, String)>,
-    token_manager: &'a Arc<TokenManager>,
+pub(crate) struct FeishuCommandContext<'a> {
+    pub(crate) db: &'a Arc<Database>,
+    pub(crate) credentials: &'a DashMap<i64, (String, String, String)>,
+    pub(crate) token_manager: &'a Arc<TokenManager>,
     /// ServiceContext：供 handle_help 查可用执行器列表（assemble_help_card_state 需要）。
-    ctx: &'a ServiceContext,
-    bot_id: i64,
-    chat_type: &'a str,
-    sender: &'a str,
-    channel: &'a str,
-    message_id: &'a str,
-    content: &'a str,
-    reaction_id: Option<&'a str>,
+    pub(crate) ctx: &'a ServiceContext,
+    pub(crate) bot_id: i64,
+    pub(crate) chat_type: &'a str,
+    pub(crate) sender: &'a str,
+    pub(crate) channel: &'a str,
+    pub(crate) message_id: &'a str,
+    pub(crate) content: &'a str,
+    pub(crate) reaction_id: Option<&'a str>,
 }
 
 /// 卡片 act:/ 动作（点击按钮要执行的副作用）。
@@ -77,16 +74,11 @@ pub(crate) enum CardAction {
     SetExecutor(String),
 }
 
-/// 把 started_at（ISO，如 "2026-07-11T14:04:37Z"）格式化成可读时间：取前 16 位 + T 换空格。
-/// 不引入 chrono 依赖，精度足够卡片展示。
-fn format_record_time(started_at: &str) -> String {
-    started_at.get(..16).unwrap_or(started_at).replace('T', " ")
-}
 
 /// 卡片 act 动作的执行结果（供 patch_after_action 渲染顶部提示）。
-struct ActionOutcome {
-    success: bool,
-    message: String,
+pub(crate) struct ActionOutcome {
+    pub(crate) success: bool,
+    pub(crate) message: String,
 }
 
 /// 编排器专用：handle_message 阶段函数之间传递的"消息预处理结果"。
@@ -170,7 +162,7 @@ impl FeishuListener {
         );
 
         let real_bot_open_id =
-            Self::resolve_bot_open_id(&self.bot_credentials, &self.token_manager, bot.id)
+            FeishuApiClient::resolve_bot_open_id(&self.bot_credentials, &self.token_manager, bot.id)
                 .await
                 .or(bot.bot_open_id.clone())
                 .unwrap_or_default();
@@ -229,7 +221,7 @@ impl FeishuListener {
     // 每个阶段职责单一，bool 返回值告知编排器是否终止。
     // 新增命令 / 改权限逻辑 / 改绑定逻辑时只动对应阶段，不会牵动整段流程。
     // ---------------------------------------------------------------
-    async fn handle_message(context: ListenerMessageContext<'_>, msg: &ChannelMessage) {
+    pub(crate) async fn handle_message(context: ListenerMessageContext<'_>, msg: &ChannelMessage) {
         // 入口日志：排查"消息为什么没反应"的第一线索
         tracing::info!(
             "[feishu:{}] handle_message: sender={}, bot_open_id={}, content={:?}, chat_type={:?}",
@@ -237,7 +229,7 @@ impl FeishuListener {
         );
         // 阶段 0：卡片回调处理（由飞书卡片按钮点击触发）
         if msg.chat_type.as_deref() == Some("card_callback") {
-            Self::handle_card_callback(context, msg).await;
+            CardActionHandler::handle_card_callback(context, msg).await;
             return;
         }
         // 阶段 0a：跳过机器人自己发的消息（不持久化、不加 reaction）
@@ -333,7 +325,7 @@ impl FeishuListener {
         bot_id: i64,
         message_id: &str,
     ) -> Option<String> {
-        Self::add_reaction(credentials, token_manager, bot_id, message_id, "THUMBSUP").await
+        FeishuApiClient::add_reaction(credentials, token_manager, bot_id, message_id, "THUMBSUP").await
     }
 
     /// 阶段 2：内置斜杠命令路由（命中并处理后返回 true）
@@ -357,14 +349,14 @@ impl FeishuListener {
             content: prep.content,
             reaction_id: prep.reaction_id.as_deref(),
         };
-        if prep.content == "/feishupush" { Self::handle_feishupush(mk_ctx()).await; return true; }
-        if prep.content == "/list" { Self::handle_list(mk_ctx()).await; return true; }
+        if prep.content == "/feishupush" { SlashCommandHandler::handle_feishupush(mk_ctx()).await; return true; }
+        if prep.content == "/list" { SlashCommandHandler::handle_list(mk_ctx()).await; return true; }
         if prep.content == "/help" || prep.content.starts_with("/help ") {
-            Self::handle_help(mk_ctx()).await; return true;
+            SlashCommandHandler::handle_help(mk_ctx()).await; return true;
         }
-        if prep.content == "/new" { Self::handle_new(mk_ctx()).await; return true; }
+        if prep.content == "/new" { SlashCommandHandler::handle_new(mk_ctx()).await; return true; }
         if prep.content == "/stop" {
-            Self::handle_stop(context.task_manager, mk_ctx()).await; return true;
+            SlashCommandHandler::handle_stop(context.task_manager, mk_ctx()).await; return true;
         }
         false
     }
@@ -438,7 +430,7 @@ impl FeishuListener {
         reaction_id: Option<&str>,
     ) {
         let Some(rid) = reaction_id else { return };
-        Self::delete_reaction(
+        FeishuApiClient::delete_reaction(
             context.credentials, context.token_manager, context.bot_id, &message.id, rid,
         ).await;
     }
@@ -834,1340 +826,43 @@ impl FeishuListener {
         Some(SlashCommandMatch { command, body })
     }
 
-    /// Handle /feishupush - cycle push level: disabled -> result_only -> all -> disabled.
-    async fn handle_feishupush(context: FeishuCommandContext<'_>) {
-        let FeishuCommandContext {
-            db,
-            credentials,
-            token_manager,
-            bot_id,
-            chat_type,
-            sender,
-            channel,
-            message_id,
-            reaction_id,
-            ..
-        } = context;
-        let (receive_id, receive_id_type) = match chat_type {
-            "p2p" => (sender.to_string(), "open_id"),
-            _ => (channel.to_string(), "chat_id"),
-        };
-
-        let target = db.get_feishu_push_target(bot_id).await.ok().flatten();
-        let current_level = target
-            .as_ref()
-            .map(|t| t.push_level.as_str())
-            .unwrap_or("disabled");
-        let new_level = match current_level {
-            "disabled" => "result_only",
-            "result_only" => "all",
-            "all" => "disabled",
-            _ => "disabled",
-        };
-
-        if let Err(e) = db.update_feishu_push_level(bot_id, new_level).await {
-            tracing::error!("[feishu:{}] /feishupush update failed: {e}", bot_id);
-            let msg = "⚠️ 操作失败，请稍后重试";
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                msg,
-            )
-            .await;
-        } else {
-            let (status_text, status_emoji) = match new_level {
-                "disabled" => ("关闭", "ℹ️"),
-                "result_only" => ("已切换为仅结论", "✅"),
-                "all" => ("已切换为全部", "✅"),
-                _ => ("未知", "⚠️"),
-            };
-            let msg = format!("{} 推送{}。", status_emoji, status_text);
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                &msg,
-            )
-            .await;
-            tracing::info!(
-                "[feishu:{}] /feishupush: push level changed to {} for bot_id={}",
-                bot_id,
-                new_level,
-                bot_id
-            );
-        }
-
-        if let Some(rid) = reaction_id {
-            Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-        }
-    }
-
-    /// Handle /list — list all registered project directories.
-    async fn handle_list(context: FeishuCommandContext<'_>) {
-        let FeishuCommandContext {
-            db,
-            credentials,
-            token_manager,
-            bot_id,
-            chat_type,
-            sender,
-            channel,
-            message_id,
-            reaction_id,
-            ..
-        } = context;
-        let (receive_id, receive_id_type) = match chat_type {
-            "p2p" => (sender.to_string(), "open_id"),
-            _ => (channel.to_string(), "chat_id"),
-        };
-
-        let directories = db.get_project_directories().await.unwrap_or_default();
-        if directories.is_empty() {
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                "📂 暂无已注册的项目目录。\n\n请在 Web 设置页「项目目录」中添加，或使用 /bind <名称> 绑定一个项目（首次使用会自动创建）。",
-            )
-            .await;
-        } else {
-            let mut lines: Vec<String> = directories
-                .iter()
-                .map(|d| {
-                    let name = d.name.as_deref().unwrap_or("(未命名)");
-                    format!("• {}  →  {}", name, d.path)
-                })
-                .collect();
-            lines.insert(0, format!("📂 已注册的项目目录（共 {} 个）：", directories.len()));
-            lines.push(String::new());
-            lines.push("💡 使用 /bind <名称> 绑定到本项目聊天".to_string());
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                &lines.join("\n"),
-            )
-            .await;
-        }
-
-        if let Some(rid) = reaction_id {
-            Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-        }
-    }
-
-    /// Handle /new — start a fresh session without resuming the previous one.
-    /// 全局内置斜杠命令，用于清空当前会话的 session，开启全新对话。
-    ///
-    /// 支持两种场景：
-    /// 1. 项目绑定场景：清除绑定的 todo/loop 会话
-    /// 2. 私聊默认响应执行器场景：清除默认执行器的会话
-    async fn handle_new(context: FeishuCommandContext<'_>) {
-        let FeishuCommandContext {
-            db,
-            credentials,
-            token_manager,
-            bot_id,
-            chat_type,
-            sender,
-            channel,
-            message_id,
-            reaction_id,
-            ..
-        } = context;
-        let (receive_id, receive_id_type) = match chat_type {
-            "p2p" => (sender.to_string(), "open_id"),
-            _ => (channel.to_string(), "chat_id"),
-        };
-
-        // 先尝试项目绑定场景
-        match db.get_feishu_project_binding(bot_id, channel).await {
-            Ok(Some(binding)) => {
-                // 清除 session_id 和 latest_record_id，使下一条消息无法 resume
-                // should_resume 的判断依赖 latest_record.session_id.is_some()，
-                // 清除后 latest_record_id=None → latest_record=None → should_resume=false
-                if let Err(e) = db.clear_feishu_binding_session(binding.id).await {
-                    tracing::error!("[feishu:{}] /new clear session failed: {e}", bot_id);
-                    Self::send_text(
-                        credentials,
-                        token_manager,
-                        bot_id,
-                        &receive_id,
-                        receive_id_type,
-                        "⚠️ 清除会话失败，请稍后重试。",
-                    )
-                    .await;
-                    if let Some(rid) = reaction_id {
-                        Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                    }
-                    return;
-                }
-
-                tracing::info!(
-                    "[feishu:{}] /new command: cleared session for binding {}, next message will start fresh",
-                    bot_id,
-                    binding.id
-                );
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "🆕 已开启新会话。\n\n发送你的任务，我将使用全新 session 执行，不再resume之前的对话。",
-                )
-                .await;
-
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-            Ok(None) => {
-                // 没有绑定项目，尝试私聊默认响应执行器场景
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] /new query binding failed: {e}", bot_id);
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "⚠️ 查询绑定失败，请稍后重试。",
-                )
-                .await;
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-        }
-
-        // 私聊默认响应执行器场景：获取 workspace 和默认执行器配置
-        let workspace_id = match db.get_agent_bot_workspace_id(bot_id).await {
-            Ok(Some(wid)) => wid,
-            Ok(None) => {
-                tracing::warn!("[feishu:{}] /new: bot has no workspace", bot_id);
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "⚠️ 未找到工作空间，无法使用 /new。",
-                )
-                .await;
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] /new query workspace failed: {e}", bot_id);
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "⚠️ 查询工作空间失败，请稍后重试。",
-                )
-                .await;
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-        };
-
-        // 获取 workspace 设置，判断默认响应类型
-        let settings = match crate::db::workspace_setting::get_workspace_settings(db, workspace_id).await {
-            Ok(Some(s)) => s,
-            Ok(None) => {
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "📭 当前未配置默认响应，无法使用 /new。",
-                )
-                .await;
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] /new query workspace settings failed: {e}", bot_id);
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "⚠️ 查询工作空间设置失败，请稍后重试。",
-                )
-                .await;
-                if let Some(rid) = reaction_id {
-                    Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-                }
-                return;
-            }
-        };
-
-        // 只处理 executor 类型的默认响应
-        if settings.default_response_type != "executor" {
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                "📭 当前默认响应类型不是执行器，无法使用 /new 清空会话。",
-            )
-            .await;
-            if let Some(rid) = reaction_id {
-                Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-            }
-            return;
-        }
-
-        let executor_name = settings.default_response_executor
-            .unwrap_or_else(|| "claudecode".to_string());
-
-        // 清空执行器 session：设置为 None
-        if let Err(e) = db.set_executor_session(workspace_id, &executor_name, None).await {
-            tracing::error!("[feishu:{}] /new clear executor session failed: {e}", bot_id);
-            Self::send_text(
-                credentials,
-                token_manager,
-                bot_id,
-                &receive_id,
-                receive_id_type,
-                "⚠️ 清除执行器会话失败，请稍后重试。",
-            )
-            .await;
-            if let Some(rid) = reaction_id {
-                Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-            }
-            return;
-        }
-
-        tracing::info!(
-            "[feishu:{}] /new command: cleared executor session for {}, workspace={}",
-            bot_id,
-            executor_name,
-            workspace_id
-        );
-        Self::send_text(
-            credentials,
-            token_manager,
-            bot_id,
-            &receive_id,
-            receive_id_type,
-            &format!("🆕 已开启新会话。\n\n下次对话将使用全新的 {} 会话，不再接续之前的对话。", executor_name),
-        )
-        .await;
-
-        if let Some(rid) = reaction_id {
-            Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-        }
-    }
-
-    /// Handle /stop — stop the currently running execution for this binding.
-    /// 与前端「停止」按钮逻辑相同：通过 task_manager 取消任务。
-    async fn handle_stop(
-        task_manager: &Arc<TaskManager>,
-        context: FeishuCommandContext<'_>,
-    ) {
-        let FeishuCommandContext {
-            db,
-            credentials,
-            token_manager,
-            bot_id,
-            chat_type,
-            sender,
-            channel,
-            message_id,
-            reaction_id,
-            ..
-        } = context;
-        let (receive_id, receive_id_type) = match chat_type {
-            "p2p" => (sender.to_string(), "open_id"),
-            _ => (channel.to_string(), "chat_id"),
-        };
-
-        match db.get_feishu_project_binding(bot_id, channel).await {
-            Ok(Some(binding)) => {
-                // 获取当前 binding 的最新执行记录
-                if let Some(record_id) = binding.latest_record_id {
-                    match db.get_execution_record(record_id).await {
-                        Ok(Some(record)) => {
-                            if record.status == crate::models::ExecutionStatus::Running {
-                                // 任务正在运行，尝试停止
-                                if let Some(ref task_id) = record.task_id {
-                                    let cancelled = task_manager.cancel(task_id).await;
-                                    if cancelled {
-                                        tracing::info!(
-                                            "[feishu:{}] /stop: cancelled task {} for record {}",
-                                            bot_id,
-                                            task_id,
-                                            record_id
-                                        );
-                                        Self::send_text(
-                                            credentials,
-                                            token_manager,
-                                            bot_id,
-                                            &receive_id,
-                                            receive_id_type,
-                                            "⏹️ 已发送停止信号，任务即将终止。",
-                                        )
-                                        .await;
-                                    } else {
-                                        // 任务不在 task_manager 中（可能已崩溃），强制更新 DB
-                                        tracing::warn!(
-                                            "[feishu:{}] /stop: task {} not in task_manager, forcing DB update",
-                                            bot_id,
-                                            task_id
-                                        );
-                                        let _ = db.force_fail_execution_record(record_id).await;
-                                        Self::send_text(
-                                            credentials,
-                                            token_manager,
-                                            bot_id,
-                                            &receive_id,
-                                            receive_id_type,
-                                            "⚠️ 任务已不在运行中（可能已异常退出），已更新状态。",
-                                        )
-                                        .await;
-                                    }
-                                } else {
-                                    Self::send_text(
-                                        credentials,
-                                        token_manager,
-                                        bot_id,
-                                        &receive_id,
-                                        receive_id_type,
-                                        "⚠️ 该执行记录没有 task_id，无法停止。",
-                                    )
-                                    .await;
-                                }
-                            } else {
-                                Self::send_text(
-                                    credentials,
-                                    token_manager,
-                                    bot_id,
-                                    &receive_id,
-                                    receive_id_type,
-                                    "ℹ️ 当前没有正在执行的任务。",
-                                )
-                                .await;
-                            }
-                        }
-                        Ok(None) => {
-                            Self::send_text(
-                                credentials,
-                                token_manager,
-                                bot_id,
-                                &receive_id,
-                                receive_id_type,
-                                "⚠️ 执行记录不存在。",
-                            )
-                            .await;
-                        }
-                        Err(e) => {
-                            tracing::error!("[feishu:{}] /stop query record failed: {e}", bot_id);
-                            Self::send_text(
-                                credentials,
-                                token_manager,
-                                bot_id,
-                                &receive_id,
-                                receive_id_type,
-                                "⚠️ 查询执行记录失败，请稍后重试。",
-                            )
-                            .await;
-                        }
-                    }
-                } else {
-                    Self::send_text(
-                        credentials,
-                        token_manager,
-                        bot_id,
-                        &receive_id,
-                        receive_id_type,
-                        "ℹ️ 当前没有执行记录可停止。",
-                    )
-                    .await;
-                }
-            }
-            Ok(None) => {
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "📭 当前聊天未绑定任何项目，无可停止的任务。",
-                )
-                .await;
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] /stop query binding failed: {e}", bot_id);
-                Self::send_text(
-                    credentials,
-                    token_manager,
-                    bot_id,
-                    &receive_id,
-                    receive_id_type,
-                    "⚠️ 查询绑定失败，请稍后重试。",
-                )
-                .await;
-            }
-        }
-
-        if let Some(rid) = reaction_id {
-            Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-        }
-    }
-
-    /// Handle /help — show interactive help card with tabbed navigation.
-    /// 点击 Tab 按钮会触发 card.action.trigger 事件，由飞书平台回调处理。
-    async fn handle_help(context: FeishuCommandContext<'_>) {
-        let FeishuCommandContext {
-            db,
-            credentials,
-            token_manager,
-            ctx,
-            bot_id,
-            chat_type: _,
-            sender,
-            message_id,
-            content,
-            reaction_id,
-            ..
-        } = context;
-
-        // 解析当前分组，默认 "status"（状态页是控制台首页）
-        let parsed = content.strip_prefix("/help ").unwrap_or("").trim().to_lowercase();
-        let group = if parsed.is_empty() { "status".to_string() } else { parsed };
-
-        // 状态感知控制台：查当前绑定/运行状态/推送级别/最近任务后渲染
-        let state = Self::assemble_help_card_state(ctx, db, bot_id, &group, 1).await;
-        let card = build_help_console_card(&state);
-        let card_json = render_card(&card, &format!("feishu:{}", sender));
-
-        // 发送卡片（reply API），失败降级纯文本
-        if let Err(e) = Self::reply_card(credentials, token_manager, bot_id, message_id, &card_json).await {
-            tracing::error!("[feishu:{}] /help send card failed: {}", bot_id, e);
-            Self::send_text(credentials, token_manager, bot_id, sender, "open_id", "📋 NTD 控制台\n\n发送 /help 打开任务控制台。").await;
-        }
-
-        if let Some(rid) = reaction_id {
-            Self::delete_reaction(credentials, token_manager, bot_id, message_id, rid).await;
-        }
-    }
-
-    /// 处理飞书卡片按钮点击回调。
-    /// card_callback 消息的 content 是 action value，按前缀分三种处理：
-    /// - `nav:/help <group>`：原地 patch 原卡片，切到对应分组；
-    /// - `cmd:/<command>`：转成命令文本，复用 handle_message 分发链路执行，
-    ///   等价于点击者在会话里发送了 `/<command>`；
-    /// - `act:/<action>`：执行动作 + patch 刷新控制台。
-    async fn handle_card_callback(context: ListenerMessageContext<'_>, msg: &ChannelMessage) {
-        let action = msg.content.trim();
-
-        // nav: 前缀 - 原地 patch 刷新控制台/历史页，拦截后直接返回。
-        if Self::handle_nav_action(&context, msg, action).await {
-            return;
-        }
-
-        // cmd: 前缀 - 把卡片点击转成命令执行。
-        // 构造一条虚拟命令消息复用 handle_message 的完整分发链路（内置命令 try_route_builtin_command
-        // + 自定义规则 route_slash_or_default_response），与用户在会话里手动发送该命令效果一致。
-        if let Some(cmd_text) = Self::parse_card_command(action) {
-            tracing::info!(
-                "[feishu:{}] card cmd → redispatch as message: {:?}",
-                context.bot_id, cmd_text
-            );
-            // chat_type 改成 p2p：避免 handle_message 又把这条消息当作 card_callback 递归处理；
-            // sender/channel/id 沿用卡片回调，让命令处理函数的回复落到原会话、指向点击者。
-            let mut cmd_msg = msg.clone();
-            cmd_msg.content = cmd_text;
-            cmd_msg.chat_type = Some("p2p".to_string());
-            // handle_message → handle_card_callback → handle_message 是静态递归，
-            // async fn 递归必须 Box::pin 引入间接层，否则 future 大小无限无法编译。
-            // 运行时 cmd_msg.chat_type 已是 p2p，不会再进 card_callback 分支，实际只递归一层。
-            Box::pin(Self::handle_message(context, &cmd_msg)).await;
-            return;
-        }
-
-        // act: 前缀 - 执行动作（新会话/停止/设推送/切绑定/解绑）+ patch 刷新控制台。
-        // 解析失败（未知 verb）落到下面的 unknown warn。
-        if let Some(parsed) = Self::parse_card_action(action) {
-            Self::execute_card_action(context, msg, parsed).await;
-            return;
-        }
-
-        tracing::warn!("[feishu:{}] unknown card action: {}", context.bot_id, action);
-    }
-
-    /// 处理 nav: 前缀的卡片动作，返回 true 表示已处理。
-    /// nav 动作是只读 patch 刷新，不产生副作用。
-    async fn handle_nav_action(context: &ListenerMessageContext<'_>, msg: &ChannelMessage, action: &str) -> bool {
-        // nav:/help <group> 重查最新状态后刷控制台（运行状态可能已变）。
-        if let Some(group) = action.strip_prefix("nav:/help ") {
-            let group_key = group.trim().to_lowercase();
-            let state = Self::assemble_help_card_state(context.ctx, context.db, context.bot_id, &group_key, 1).await;
-            let card = build_help_console_card(&state);
-            Self::patch_rendered_card(context, msg, &card).await;
-            return true;
-        }
-        // nav:/history [page] - 分页查看执行历史。
-        if let Some(page_arg) = action.strip_prefix("nav:/history") {
-            let page = page_arg.trim().parse::<usize>().unwrap_or(1).max(1);
-            Self::patch_history_page(context, msg, page).await;
-            return true;
-        }
-        // nav:/todos <page> - 事项分页（每页 10）。
-        if let Some(page_arg) = action.strip_prefix("nav:/todos") {
-            let page = page_arg.trim().parse::<usize>().unwrap_or(1).max(1);
-            let state = Self::assemble_help_card_state(context.ctx, context.db, context.bot_id, "todo", page).await;
-            let card = build_help_console_card(&state);
-            Self::patch_rendered_card(context, msg, &card).await;
-            return true;
-        }
-        // nav:/loops <page> - 环路分页（每页 10）。
-        if let Some(page_arg) = action.strip_prefix("nav:/loops") {
-            let page = page_arg.trim().parse::<usize>().unwrap_or(1).max(1);
-            let state = Self::assemble_help_card_state(context.ctx, context.db, context.bot_id, "loop", page).await;
-            let card = build_help_console_card(&state);
-            Self::patch_rendered_card(context, msg, &card).await;
-            return true;
-        }
-        false
-    }
-
-    /// 执行卡片 act 动作，执行后 patch 刷新控制台。
-    async fn execute_card_action(context: ListenerMessageContext<'_>, msg: &ChannelMessage, action: CardAction) {
-        let outcome = Self::run_card_action(&context, msg, &action).await;
-        let group = Self::action_target_group(&action);
-        Self::patch_after_action(&context, msg, group, &outcome).await;
-    }
-
-    /// 按 CardAction 分发到具体执行函数。
-    async fn run_card_action(
-        context: &ListenerMessageContext<'_>,
-        msg: &ChannelMessage,
-        action: &CardAction,
-    ) -> ActionOutcome {
-        match action {
-            CardAction::Push(level) => Self::act_push(context, level).await,
-            CardAction::New => Self::act_new(context).await,
-            CardAction::Stop => Self::act_stop(context).await,
-            CardAction::Bind(workspace_id) => Self::act_bind(context, *workspace_id).await,
-            CardAction::RunTodo(todo_id) => Self::act_run_todo(context, msg, *todo_id).await,
-            CardAction::RunLoop(loop_id) => Self::act_run_loop(context, msg, *loop_id).await,
-            CardAction::SetExecutor(name) => Self::act_set_executor(context, name).await,
-        }
-    }
-
-    /// act 动作执行后刷新到的目标 Tab。
-    fn action_target_group(action: &CardAction) -> &'static str {
-        match action {
-            CardAction::Bind(_) | CardAction::Push(_) | CardAction::SetExecutor(_) => "workspace",
-            CardAction::RunTodo(_) => "todo",
-            CardAction::RunLoop(_) => "loop",
-            _ => "status",
-        }
-    }
-
-    /// bot 所属 workspace 的默认执行器（如 dev 的 pi）。
-    async fn workspace_default_executor(db: &Database, bot_id: i64) -> Option<String> {
-        let wid = db.get_agent_bot_workspace_id(bot_id).await.ok().flatten()?;
-        let settings = crate::db::workspace_setting::get_workspace_settings(db, wid).await.ok().flatten()?;
-        settings.default_response_executor
-    }
-
-    /// auto-seed：确保 workspace 的 default_response_type=executor。
-    /// 移除 binding 路径后 chat 全走 default_response，只 executor 分支可靠回复；切换 workspace 时兜底。
-    async fn ensure_default_response_executor(db: &Database, workspace_id: i64) {
-        let existing = crate::db::workspace_setting::get_workspace_settings(db, workspace_id).await.ok().flatten();
-        let need_seed = existing.as_ref().map(|s| s.default_response_type != "executor").unwrap_or(true);
-        if need_seed {
-            // executor 用该 workspace 已配的（若有），否则 None（dispatch 时兜底 claudecode）
-            let executor = existing.and_then(|s| s.default_response_executor);
-            let _ = crate::db::workspace_setting::upsert_workspace_settings(
-                db,
-                workspace_id,
-                Some("executor".to_string()),
-                None,
-                None,
-                executor,
-                // 飞书 listener 仅切 default_response，不动 workspace 共识 prompt
-                None,
-            )
-            .await;
-        }
-    }
-
-    /// 设置推送级别（直接设值，不走 /feishupush 循环）。
-    async fn act_push(context: &ListenerMessageContext<'_>, level: &str) -> ActionOutcome {
-        match context.db.update_feishu_push_level(context.bot_id, level).await {
-            Ok(_) => ActionOutcome { success: true, message: format!("推送级别已更新为 {level}") },
-            Err(e) => ActionOutcome { success: false, message: format!("设置失败：{e}") },
-        }
-    }
-
-    /// 开启新会话：清当前 workspace 默认执行器的 session。
-    async fn act_new(context: &ListenerMessageContext<'_>) -> ActionOutcome {
-        let Some(wid) = context.db.get_agent_bot_workspace_id(context.bot_id).await.ok().flatten() else {
-            return ActionOutcome { success: false, message: "未设置工作空间".to_string() };
-        };
-        let executor = Self::workspace_default_executor(context.db, context.bot_id)
-            .await
-            .unwrap_or_else(|| "claudecode".to_string());
-        match context.db.set_executor_session(wid, &executor, None).await {
-            Ok(_) => ActionOutcome { success: true, message: "已开启新会话".to_string() },
-            Err(e) => ActionOutcome { success: false, message: format!("失败：{e}") },
-        }
-    }
-
-    /// 停止当前 workspace 的运行任务（by workspace + ExecutionStatus::Running 直接查）。
-    async fn act_stop(context: &ListenerMessageContext<'_>) -> ActionOutcome {
-        let Some(wid) = context.db.get_agent_bot_workspace_id(context.bot_id).await.ok().flatten() else {
-            return ActionOutcome { success: false, message: "未设置工作空间".to_string() };
-        };
-        // 直接按 workspace 查运行中的记录，不依赖最近 N 条（避免旧记录淹没了 running 记录）
-        let Ok(records) = context.db.get_running_records_by_workspace(wid).await else {
-            return ActionOutcome { success: false, message: "查询失败".to_string() };
-        };
-        let Some(running) = records.into_iter().next() else {
-            return ActionOutcome { success: false, message: "没有运行中的任务".to_string() };
-        };
-        let Some(task_id) = running.task_id.as_deref() else {
-            return ActionOutcome { success: false, message: "任务缺少 task_id".to_string() };
-        };
-        if context.task_manager.cancel(task_id).await {
-            ActionOutcome { success: true, message: "已发送停止信号，任务即将终止".to_string() }
-        } else {
-            let _ = context.db.force_fail_execution_record(running.id).await;
-            ActionOutcome { success: true, message: "任务未在运行，已强制标记结束".to_string() }
-        }
-    }
-
-    /// 切换工作空间：级联清旧 binding + 改 agent_bot.workspace_id + auto-seed default_response。
-    async fn act_bind(context: &ListenerMessageContext<'_>, workspace_id: i64) -> ActionOutcome {
-        let bot_id = context.bot_id;
-        // 级联（对齐 move_bot_to_workspace）：删 pending binding / disable 旧 binding
-        if let Ok(bindings) = context.db.get_feishu_project_bindings(bot_id).await {
-            for b in bindings {
-                if b.chat_id == crate::models::PENDING_CHAT_ID {
-                    let _ = context.db.delete_feishu_project_binding(b.id).await;
-                } else {
-                    let _ = context.db.update_feishu_project_binding_enabled(b.id, false).await;
-                }
-            }
-        }
-        if let Err(e) = context.db.update_agent_bot_workspace_id(bot_id, workspace_id).await {
-            return ActionOutcome { success: false, message: format!("切换工作空间失败：{e}") };
-        }
-        // auto-seed default_response_type=executor，确保切完后 chat 消息有回复
-        Self::ensure_default_response_executor(context.db, workspace_id).await;
-        let name = context
-            .db
-            .get_workspace_name_by_id(workspace_id)
-            .await
-            .ok()
-            .flatten()
-            .unwrap_or_else(|| format!("#{workspace_id}"));
-        ActionOutcome { success: true, message: format!("已切换到工作空间「{name}」") }
-    }
-
-    /// 触发事项：后台跑该 todo，结果通过 ExecEvent::Finished → FeishuPushService 推回当前 chat。
-    async fn act_run_todo(context: &ListenerMessageContext<'_>, msg: &ChannelMessage, todo_id: i64) -> ActionOutcome {
-        use crate::executor_service::{run_todo_execution, RunTodoExecutionRequest};
-        let (receive_id, receive_id_type) = Self::resolve_receive_target(msg);
-        let workspace_id = context.db.get_agent_bot_workspace_id(context.bot_id).await.ok().flatten();
-        let todo = match context.db.get_todo(todo_id).await {
-            Ok(Some(t)) => t,
-            _ => return ActionOutcome { success: false, message: format!("事项 #{todo_id} 不存在") },
-        };
-        // 校验 todo 的 workspace_id 与 bot 当前 workspace 一致，防止旧卡片跨 workspace 执行
-        if todo.workspace_id != workspace_id {
-            return ActionOutcome {
-                success: false,
-                message: format!("事项 #{todo_id} 不属于当前工作空间，无法执行"),
-            };
-        }
-        let title = todo.title.clone();
-        let req = RunTodoExecutionRequest {
-            db: context.db.clone(),
-            executor_registry: context.ctx.executor_registry.clone(),
-            tx: context.ctx.tx.clone(),
-            task_manager: context.task_manager.clone(),
-            config: context.ctx.config.clone(),
-            todo_id,
-            message: todo.prompt,
-            req_executor: todo.executor,
-            req_model: None,
-            trigger_type: "feishu_card".to_string(),
-            params: None,
-            resume_session_id: None,
-            resume_message: None,
-            source_todo_id: None,
-            source_todo_title: None,
-            loop_step_execution_id: None,
-            step_id: None,
-            feishu_bot_id: Some(context.bot_id),
-            feishu_receive_id: Some(receive_id.to_string()),
-            feishu_receive_id_type: Some(receive_id_type.to_string()),
-            workspace_path: None,
-            workspace_id,
-            // 飞书卡片触发路径：注入专家上下文，卡片触发也需尊重 todo 的专家绑定
-            expert_manager: Some(context.ctx.expert_manager.clone()),
-        };
-        // fire-and-forget：后台执行不阻塞卡片 patch；结果由推送通道发回 chat
-        tokio::spawn(async move {
-            let _ = run_todo_execution(req).await;
-        });
-        ActionOutcome { success: true, message: format!("已触发事项「{title}」") }
-    }
-
-    /// 触发环路：LoopRunner::spawn_run 后台执行，整环结束推回 chat。
-    async fn act_run_loop(context: &ListenerMessageContext<'_>, msg: &ChannelMessage, loop_id: i64) -> ActionOutcome {
-        let Some(runner) = context.debounce.loop_runner() else {
-            return ActionOutcome { success: false, message: "环路执行器未就绪".to_string() };
-        };
-        let (receive_id, receive_id_type) = Self::resolve_receive_target(msg);
-        let loop_ = match context.db.get_loop(loop_id).await {
-            Ok(Some(l)) => l,
-            _ => return ActionOutcome { success: false, message: format!("环路 #{loop_id} 不存在") },
-        };
-        // 校验 loop 的 workspace_id 与 bot 当前 workspace 一致，防止旧卡片跨 workspace 执行
-        let workspace_id = context.db.get_agent_bot_workspace_id(context.bot_id).await.ok().flatten();
-        if loop_.workspace_id != workspace_id {
-            return ActionOutcome {
-                success: false,
-                message: format!("环路 #{loop_id} 不属于当前工作空间，无法执行"),
-            };
-        }
-        let name = loop_.name;
-        runner.clone().spawn_run(
-            loop_id,
-            None,
-            "feishu_card",
-            serde_json::json!({}),
-            Some(context.bot_id),
-            Some(receive_id.to_string()),
-            Some(receive_id_type.to_string()),
-        );
-        ActionOutcome { success: true, message: format!("已触发环路「{name}」") }
-    }
-
-    /// 把当前 workspace 的「默认响应执行器」设为指定 executor：
-    /// 既写 workspace_settings.default_response_executor，又把 default_response_type 改为 "executor"。
-    /// 只写 executor 字段而不改 type 是有 bug 的——dispatch_default_response 按 default_response_type
-    /// 分发，type 仍是 todo/loop 时点了「默认响应执行器」按钮也不会真的切到执行器，所以两字段必须同改。
-    /// executor 名必须是已注册的（ExecutorType::as_str），否则视为无效拒绝写入。
-    async fn act_set_executor(context: &ListenerMessageContext<'_>, executor_name: &str) -> ActionOutcome {
-        let Some(wid) = context.db.get_agent_bot_workspace_id(context.bot_id).await.ok().flatten() else {
-            return ActionOutcome { success: false, message: "未设置工作空间".to_string() };
-        };
-        // 校验 executor 已注册，避免把无效名写进 settings 让下次 dispatch 失败
-        let registered: Vec<String> = Self::registered_executor_names(context).await;
-        if !registered.iter().any(|s| s == executor_name) {
-            return ActionOutcome {
-                success: false,
-                message: format!("执行器 {executor_name} 未注册（可用：{}）", registered.join(", ")),
-            };
-        }
-        // type 与 executor 同改：type=executor 让 dispatch 走执行器分支，executor 字段指定具体执行器。
-        // todo_id/loop_id 传 None 表示不动旧值——切回 todo/loop 时这些旧值还有用，不在这里清。
-        match crate::db::workspace_setting::upsert_workspace_settings(
-            context.db,
-            wid,
-            Some("executor".to_string()),
-            None,
-            None,
-            Some(executor_name.to_string()),
-            // 仅切默认执行器，不动 workspace 共识 prompt
-            None,
-        )
-        .await
-        {
-            Ok(_) => ActionOutcome { success: true, message: format!("默认响应执行器已设为 {executor_name}") },
-            Err(e) => ActionOutcome { success: false, message: format!("设置失败：{e}") },
-        }
-    }
-
-    /// 从 executor_registry 拉出所有已注册执行器名（ExecutorType::as_str），返回 Vec<String>。
-    /// 抽出来一是让 act_set_executor 更短，二是注册名查询本身也是可复用的小步骤。
-    async fn registered_executor_names(context: &ListenerMessageContext<'_>) -> Vec<String> {
-        context
-            .ctx
-            .executor_registry
-            .list_executors()
-            .await
-            .into_iter()
-            .map(|t| t.as_str().to_string())
-            .collect()
-    }
-
-    /// act 执行后 patch 刷新控制台：assemble 最新状态 + 顶部插入操作结果提示。
-    async fn patch_after_action(
-        context: &ListenerMessageContext<'_>,
-        msg: &ChannelMessage,
-        group: &str,
-        outcome: &ActionOutcome,
-    ) {
-        let state = Self::assemble_help_card_state(context.ctx, context.db, context.bot_id, group, 1).await;
-        let mut card = build_help_console_card(&state);
-        let icon = if outcome.success { "✅" } else { "⚠️" };
-        let tip = CardElement::Markdown(CardMarkdown { content: format!("{icon} {}", outcome.message) });
-        card.elements.insert(0, tip);
-        Self::patch_rendered_card(context, msg, &card).await;
-    }
-
-    /// 渲染卡片并 patch 到原消息（nav/act 刷新共用）。
-    async fn patch_rendered_card(context: &ListenerMessageContext<'_>, msg: &ChannelMessage, card: &Card) {
-        let session_key = format!("feishu:{}", msg.sender);
-        let card_json = render_card(card, &session_key);
-        if let Err(e) = Self::patch_card(context.credentials, context.token_manager, context.bot_id, &msg.id, &card_json).await {
-            tracing::warn!("[feishu:{}] patch card failed: {e}", context.bot_id);
-        }
-    }
-
-    /// 历史子页：按当前 workspace 分页查执行记录后 patch。
-    async fn patch_history_page(context: &ListenerMessageContext<'_>, msg: &ChannelMessage, page: usize) {
-        const PER_PAGE: i64 = 10;
-        let offset = page.saturating_sub(1) as i64 * PER_PAGE;
-        let (items, total) = Self::query_history(context.db, context.bot_id, PER_PAGE, offset).await;
-        let total_pages = (total.max(0) as usize).div_ceil(PER_PAGE as usize);
-        let card = build_history_card(&items, page, total_pages.max(1));
-        Self::patch_rendered_card(context, msg, &card).await;
-    }
-
-    /// 按 bot 的 workspace 分页查执行记录 → HistoryItem + 总数。
-    async fn query_history(db: &Database, bot_id: i64, limit: i64, offset: i64) -> (Vec<HistoryItem>, i64) {
-        let Some(wid) = db.get_agent_bot_workspace_id(bot_id).await.ok().flatten() else {
-            return (vec![], 0);
-        };
-        match db.get_execution_records_by_workspace(wid, limit, offset).await {
-            Ok((records, total)) => (records.into_iter().map(Self::record_to_history_item).collect(), total),
-            Err(_) => (vec![], 0),
-        }
-    }
-
-    /// ExecutionRecord → 历史子页列表项（状态 emoji + 标题 + 触发类型 + 时间）。
-    fn record_to_history_item(r: crate::models::ExecutionRecord) -> HistoryItem {
-        use crate::models::ExecutionStatus;
-        let status_icon = match r.status {
-            ExecutionStatus::Success => "✅",
-            ExecutionStatus::Running => "⏳",
-            ExecutionStatus::Failed => "❌",
-        };
-        HistoryItem {
-            status_icon: status_icon.to_string(),
-            title: r.source_todo_title.clone().unwrap_or_else(|| r.command.clone()),
-            trigger: r.trigger_type,
-            time_desc: format_record_time(&r.started_at),
-        }
-    }
-
-    /// 解析卡片回调 action 里的命令文本，供 handle_card_callback 的 cmd: 分支使用。
-    /// `cmd:/new` → `Some("/new")`；`cmd:/bind foo` → `Some("/bind foo")`（保留参数）；
-    /// 非 `cmd:/` 前缀（nav:/act:/未知/空）→ None。
-    /// 抽成纯函数便于单测命令文本拼装，也让 handle_card_callback 的 cmd: 分支保持简洁。
-    fn parse_card_command(action: &str) -> Option<String> {
-        action.strip_prefix("cmd:/").map(|cmd| format!("/{}", cmd))
-    }
-
-    /// 解析卡片 act:/ 动作字符串为 CardAction。
-    /// "act:/stop"→Stop；"act:/bind myapp"→Bind("myapp")；"act:/push result_only"→Push("result_only")。
-    /// bind/push 需要参数，缺参数返回 None；未知 verb 返回 None。纯函数便于单测。
-    fn parse_card_action(action: &str) -> Option<CardAction> {
-        let rest = action.strip_prefix("act:/")?;
-        // splitn(2) 让参数部分可含空格（虽然当前不会，但留余地），verb 与 arg 用首个空白分隔。
-        let mut parts = rest.splitn(2, char::is_whitespace);
-        let verb = parts.next()?.trim();
-        let arg = parts.next().map(|s| s.trim()).filter(|s| !s.is_empty());
-        Some(match verb {
-            "stop" => CardAction::Stop,
-            "new" => CardAction::New,
-            "push" => CardAction::Push(arg?.to_string()),
-            "setexecutor" => CardAction::SetExecutor(arg?.to_string()),
-            "bind" => CardAction::Bind(arg?.parse().ok()?),
-            "runtodo" => CardAction::RunTodo(arg?.parse().ok()?),
-            "runloop" => CardAction::RunLoop(arg?.parse().ok()?),
-            _ => return None,
-        })
-    }
-
-    /// 把卡片回调消息解析成回信接收者 (receive_id, receive_id_type)。
-    /// card_callback 的 chat_type 不是 p2p/group：msg.channel(chat_id)非空 → 群聊用 chat_id；
-    /// 否则回退到点击者 open_id（私聊）。供 act:/runtodo、act:/runloop 等「回信给点击者」的动作复用。
-    fn resolve_receive_target(msg: &ChannelMessage) -> (&str, &str) {
-        if !msg.channel.is_empty() {
-            (msg.channel.as_str(), "chat_id")
-        } else {
-            (msg.sender.as_str(), "open_id")
-        }
-    }
-
-    /// 组装 /help 卡片状态：按 agent_bot.workspace_id 查该 workspace 的摘要/事项/环路/最近任务 + 所有工作空间。
-    /// handle_help、nav 切页、act 执行后刷新都复用它（只读 db，运行状态取最近记录里的 running）。
-    /// ctx 用于查询已注册的执行器列表（工作空间页渲染按钮排用）。
-    async fn assemble_help_card_state(
-        ctx: &ServiceContext,
-        db: &Database,
-        bot_id: i64,
-        current_group: &str,
-        page: usize,
-    ) -> HelpCardState {
-        let wid = db.get_agent_bot_workspace_id(bot_id).await.ok().flatten();
-        let workspace = match wid {
-            Some(id) => Self::build_workspace_summary(db, id).await,
-            None => None,
-        };
-        let push_level = db
-            .get_feishu_push_target(bot_id)
-            .await
-            .ok()
-            .flatten()
-            .map(|t| t.push_level)
-            .unwrap_or_else(|| "result_only".to_string());
-        // 最近任务 + 运行状态都来自该 workspace 的最近执行记录
-        let (recent_records, is_running) = Self::recent_records_and_running(db, wid).await;
-        let todos = match wid {
-            Some(id) => db
-                // 056：卡片只展示最近 20 条摘要（id+标题+状态图标），
-                // 用 brief 接口替代整行全量拉取；take(20) 截断保持卡片体积。
-                // DB 失败时记 error（含 bot/ws 上下文）后降级为空列表——
-                // 卡片缺区块可接受，但静默吞错会让故障无法定位（CodeRabbit#4）。
-                .get_todo_briefs(Some(id), None, None)
-                .await
-                .map(|ts| ts.into_iter().take(20).map(Self::brief_to_item).collect())
-                .unwrap_or_else(|e| {
-                    tracing::error!("飞书卡片加载 todo 摘要失败（bot={bot_id}, ws={id}）: {e}");
-                    Vec::new()
-                }),
-            None => vec![],
-        };
-        let loops = match wid {
-            Some(id) => db
-                .list_loops_with_counts(Some(id))
-                .await
-                .ok()
-                .map(|ls| ls.into_iter().map(Self::loop_to_item).collect())
-                .unwrap_or_default(),
-            None => vec![],
-        };
-        let workspaces = db
-            .get_project_directories()
-            .await
-            .ok()
-            .unwrap_or_default()
-            .into_iter()
-            .map(|d| WorkspaceItem {
-                name: d.name.clone().unwrap_or_else(|| d.path.clone()),
-                id: d.id,
-                is_current: wid == Some(d.id),
-            })
-            .collect();
-        // 已注册执行器列表 + 标记当前 workspace 配的默认执行器，供工作空间页渲染按钮排
-        let current_executor = workspace.as_ref().map(|w| w.executor.as_str()).unwrap_or("");
-        let available_executors = ctx
-            .executor_registry
-            .list_executors()
-            .await
-            .into_iter()
-            .map(|t| {
-                let name = t.as_str().to_string();
-                let is_current = name == current_executor;
-                ExecutorOption { name, is_current }
-            })
-            .collect();
-        HelpCardState {
-            current_group: current_group.to_string(),
-            workspace,
-            is_running,
-            push_level,
-            recent_records,
-            todos,
-            loops,
-            workspaces,
-            page,
-            available_executors,
-        }
-    }
-
-    /// 当前 workspace 摘要（名 + 默认执行器）。
-    async fn build_workspace_summary(db: &Database, workspace_id: i64) -> Option<WorkspaceSummary> {
-        let name = db.get_workspace_name_by_id(workspace_id).await.ok().flatten()?;
-        let executor = crate::db::workspace_setting::get_workspace_settings(db, workspace_id)
-            .await
-            .ok()
-            .flatten()
-            .and_then(|s| s.default_response_executor)
-            .unwrap_or_else(|| "claudecode".to_string());
-        Some(WorkspaceSummary { id: workspace_id, name, executor })
-    }
-
-    /// 该 workspace 最近 5 条执行记录 → RecentTaskItem；顺带判断是否有 running。
-    async fn recent_records_and_running(db: &Database, wid: Option<i64>) -> (Vec<RecentTaskItem>, bool) {
-        let Some(id) = wid else {
-            return (vec![], false);
-        };
-        let Ok((records, _)) = db.get_execution_records_by_workspace(id, 5, 0).await else {
-            return (vec![], false);
-        };
-        let is_running = records.iter().any(|r| r.status == crate::models::ExecutionStatus::Running);
-        let items = records.into_iter().map(|r| Self::record_to_recent_item(&r)).collect();
-        (items, is_running)
-    }
-
-    /// Todo → 事项页列表项。
-    /// TodoBrief → 卡片「事项列表」项（056：摘要字段足够拼状态图标，无需整行 Todo）。
-    fn brief_to_item(t: crate::models::TodoBrief) -> TodoItem {
-        use crate::models::TodoStatus;
-        let status_icon = match t.status {
-            TodoStatus::Completed => "✅",
-            TodoStatus::Running | TodoStatus::InProgress => "▶️",
-            _ => "⏸️",
-        };
-        TodoItem { id: t.id, title: t.title, status_icon: status_icon.to_string() }
-    }
-
-    /// LoopListRow → 环路页列表项。
-    fn loop_to_item(l: crate::db::loop_::LoopListRow) -> LoopItem {
-        LoopItem { id: l.loop_.id, name: l.loop_.name, status: l.loop_.status }
-    }
-
-    /// ExecutionRecord → 卡片「最近任务」项（状态 emoji + 标题 + 时间）。
-    fn record_to_recent_item(r: &crate::models::ExecutionRecord) -> RecentTaskItem {
-        use crate::models::ExecutionStatus;
-        let status_icon = match r.status {
-            ExecutionStatus::Success => "✅",
-            ExecutionStatus::Running => "⏳",
-            ExecutionStatus::Failed => "❌",
-        };
-        // 标题优先用触发源标题，其次结果文本，最后命令
-        let title = r.source_todo_title.clone().or(r.result.clone()).unwrap_or_else(|| r.command.clone());
-        RecentTaskItem {
-            status_icon: status_icon.to_string(),
-            title,
-            time_desc: format_record_time(&r.started_at),
-        }
-    }
-
-    /// Patch an existing interactive card message with new content.
-    async fn patch_card(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        message_id: &str,
-        card_json: &str,
-    ) -> anyhow::Result<()> {
-        let base_url = Self::base_url(credentials, bot_id)
-            .ok_or_else(|| anyhow::anyhow!("no base_url for bot {}", bot_id))?;
-        let token = Self::get_tenant_token(credentials, token_manager, bot_id)
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no token for bot {}", bot_id))?;
-
-        let client = reqwest::Client::new();
-        let url = format!(
-            "{}/open-apis/im/v1/messages/{}",
-            base_url, message_id
-        );
-
-        let body = serde_json::json!({
-            "msg_type": "interactive",
-            "content": card_json
-        });
-
-        let res = client
-            .patch(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("patch_card request failed: {}", e))?;
-
-        let status = res.status();
-        if !status.is_success() {
-            let body: serde_json::Value = res.json().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("patch_card failed: {} {:?}", status, body));
-        }
-
-        tracing::debug!("[feishu:{}] patch_card ok for message {}", bot_id, message_id);
-        Ok(())
-    }
-
-    /// Send a plain text message to a Feishu recipient.
-    async fn send_text(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        receive_id: &str,
-        receive_id_type: &str,
-        text: &str,
-    ) {
-        let base_url = match Self::base_url(credentials, bot_id) {
-            Some(u) => u,
-            None => return,
-        };
-        let token = match Self::get_tenant_token(credentials, token_manager, bot_id).await {
-            Some(t) => t,
-            None => return,
-        };
-
-        let client = reqwest::Client::new();
-        let url = format!(
-            "{}/open-apis/im/v1/messages?receive_id_type={}",
-            base_url, receive_id_type
-        );
-        let body = serde_json::json!({
-            "receive_id": receive_id,
-            "msg_type": "text",
-            "content": serde_json::to_string(&serde_json::json!({ "text": text })).unwrap_or_default()
-        });
-
-        match client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await
-        {
-            Ok(res) => {
-                let status = res.status();
-                if !status.is_success() {
-                    tracing::error!("[feishu:{}] send_text failed: status={}", bot_id, status);
-                } else {
-                    tracing::debug!(
-                        "[feishu:{}] send_text ok to {} ({})",
-                        bot_id,
-                        receive_id,
-                        receive_id_type
-                    );
-                }
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] send_text request failed: {e}", bot_id);
-            }
-        }
-    }
-
-    /// Send an interactive card message to a Feishu recipient.
-    #[allow(dead_code)]
-    async fn send_card(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        receive_id: &str,
-        receive_id_type: &str,
-        card_json: &str,
-    ) -> anyhow::Result<()> {
-        let base_url = Self::base_url(credentials, bot_id)
-            .ok_or_else(|| anyhow::anyhow!("no base_url for bot {}", bot_id))?;
-        let token = Self::get_tenant_token(credentials, token_manager, bot_id)
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no token for bot {}", bot_id))?;
-
-        let client = reqwest::Client::new();
-        let url = format!(
-            "{}/open-apis/im/v1/messages?receive_id_type={}",
-            base_url, receive_id_type
-        );
-
-        // 飞书 Interactive Card 的 content 直接是 JSON 字符串，不需要额外的嵌套
-        let body = serde_json::json!({
-            "receive_id": receive_id,
-            "msg_type": "interactive",
-            "content": card_json
-        });
-
-        let res = client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("send_card request failed: {}", e))?;
-
-        let status = res.status();
-        if !status.is_success() {
-            let body: serde_json::Value = res.json().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("send_card failed: {} {:?}", status, body));
-        }
-
-        tracing::debug!(
-            "[feishu:{}] send_card ok to {} ({})",
-            bot_id, receive_id, receive_id_type
-        );
-        Ok(())
-    }
-
-    /// Reply to a message with an interactive card.
-    async fn reply_card(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        message_id: &str,
-        card_json: &str,
-    ) -> anyhow::Result<()> {
-        let base_url = Self::base_url(credentials, bot_id)
-            .ok_or_else(|| anyhow::anyhow!("no base_url for bot {}", bot_id))?;
-        let token = Self::get_tenant_token(credentials, token_manager, bot_id)
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no token for bot {}", bot_id))?;
-
-        let client = reqwest::Client::new();
-        // 使用 reply API 而不是 create
-        let url = format!(
-            "{}/open-apis/im/v1/messages/{}/reply",
-            base_url, message_id
-        );
-
-        let body = serde_json::json!({
-            "msg_type": "interactive",
-            "content": card_json
-        });
-
-        let res = client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await
-            .map_err(|e| anyhow::anyhow!("reply_card request failed: {}", e))?;
-
-        let status = res.status();
-        if !status.is_success() {
-            let body: serde_json::Value = res.json().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("reply_card failed: {} {:?}", status, body));
-        }
-
-        tracing::debug!(
-            "[feishu:{}] reply_card ok to message {}",
-            bot_id, message_id
-        );
-        Ok(())
-    }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
     /// Send a message via a specific bot's channel.
     pub async fn send(&self, bot_id: i64, text: &str, recipient: &str) -> anyhow::Result<()> {
@@ -2179,281 +874,15 @@ impl FeishuListener {
         }
     }
 
-    /// Send a raw text message using a specific receive_id_type.
-    pub async fn send_raw(
-        &self,
-        bot_id: i64,
-        receive_id: &str,
-        receive_id_type: &str,
-        text: &str,
-    ) -> anyhow::Result<()> {
-        let base_url = Self::base_url(&self.bot_credentials, bot_id)
-            .ok_or_else(|| anyhow::anyhow!("no credentials for bot {}", bot_id))?;
-        let token = Self::get_tenant_token(&self.bot_credentials, &self.token_manager, bot_id)
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no token for bot {}", bot_id))?;
 
-        let client = reqwest::Client::new();
-        let url = format!(
-            "{}/open-apis/im/v1/messages?receive_id_type={}",
-            base_url, receive_id_type
-        );
-        let body = serde_json::json!({
-            "receive_id": receive_id,
-            "msg_type": "text",
-            "content": serde_json::to_string(&serde_json::json!({ "text": text })).unwrap_or_default()
-        });
-
-        let res = client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await?;
-
-        let status = res.status();
-        if !status.is_success() {
-            let body: serde_json::Value = res.json().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("send_raw failed: {} {:?}", status, body));
-        }
-
-        Ok(())
-    }
-
-    /// Send a card message using a specific receive_id_type.
-    pub async fn send_card_raw(
-        &self,
-        bot_id: i64,
-        receive_id: &str,
-        receive_id_type: &str,
-        card_json: &str,
-    ) -> anyhow::Result<()> {
-        let base_url = Self::base_url(&self.bot_credentials, bot_id)
-            .ok_or_else(|| anyhow::anyhow!("no credentials for bot {}", bot_id))?;
-        let token = Self::get_tenant_token(&self.bot_credentials, &self.token_manager, bot_id)
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no token for bot {}", bot_id))?;
-
-        let client = reqwest::Client::new();
-        let url = format!(
-            "{}/open-apis/im/v1/messages?receive_id_type={}",
-            base_url, receive_id_type
-        );
-        // 飞书 API 要求 content 字段是字符串格式的 JSON
-        // json! 宏会自动将 &str 转义为 JSON 字符串值，无需手动处理
-        let body = serde_json::json!({
-            "receive_id": receive_id,
-            "msg_type": "interactive",
-            "content": card_json
-        });
-
-        let res = client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .header("Content-Type", "application/json")
-            .json(&body)
-            .send()
-            .await?;
-
-        let status = res.status();
-        if !status.is_success() {
-            let body: serde_json::Value = res.json().await.unwrap_or_default();
-            return Err(anyhow::anyhow!("send_card_raw failed: {} {:?}", status, body));
-        }
-
-        Ok(())
-    }
 
     // --- Feishu API helpers ---
 
-    fn base_url(
-        credentials: &DashMap<i64, (String, String, String)>,
-        bot_id: i64,
-    ) -> Option<String> {
-        let domain = credentials.get(&bot_id)?.2.clone();
-        Some(if domain == "lark" {
-            "https://open.larksuite.com".to_string()
-        } else {
-            "https://open.feishu.cn".to_string()
-        })
-    }
 
-    fn build_sdk_config(
-        credentials: &DashMap<i64, (String, String, String)>,
-        bot_id: i64,
-    ) -> Option<FeishuSdkConfig> {
-        let ref_val = credentials.get(&bot_id)?;
-        let (app_id, app_secret, domain) =
-            (ref_val.0.clone(), ref_val.1.clone(), ref_val.2.clone());
-        let base_url = if domain == "lark" {
-            "https://open.larksuite.com"
-        } else {
-            "https://open.feishu.cn"
-        };
 
-        Some(
-            FeishuSdkConfig::builder()
-                .app_id(app_id)
-                .app_secret(app_secret)
-                .base_url(base_url)
-                .enable_token_cache(true)
-                .http_client(reqwest::Client::new())
-                .build(),
-        )
-    }
 
-    async fn get_tenant_token(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-    ) -> Option<String> {
-        let sdk_config = Self::build_sdk_config(credentials, bot_id)?;
-        match token_manager.get_tenant_access_token(&sdk_config).await {
-            Ok(token) => Some(token),
-            Err(err) => {
-                tracing::warn!("[feishu:{}] 获取 tenant_access_token 失败: {}", bot_id, err);
-                None
-            }
-        }
-    }
 
-    async fn resolve_bot_open_id(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-    ) -> Option<String> {
-        let token = Self::get_tenant_token(credentials, token_manager, bot_id).await?;
-        let base_url = Self::base_url(credentials, bot_id)?;
 
-        let client = reqwest::Client::new();
-        let res = client
-            .get(format!("{base_url}/open-apis/bot/v3/info"))
-            .header("Authorization", format!("Bearer {token}"))
-            .send()
-            .await
-            .ok()?;
-
-        let body: serde_json::Value = res.json().await.ok()?;
-        body.get("bot")
-            .and_then(|b| b.get("open_id"))
-            .and_then(|v| v.as_str())
-            .map(String::from)
-    }
-
-    /// Add reaction, returns reaction_id on success.
-    async fn add_reaction(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        message_id: &str,
-        emoji_type: &str,
-    ) -> Option<String> {
-        let token = Self::get_tenant_token(credentials, token_manager, bot_id).await?;
-        let base_url = Self::base_url(credentials, bot_id)?;
-
-        let client = reqwest::Client::new();
-        let url = format!("{base_url}/open-apis/im/v1/messages/{message_id}/reactions");
-        let body_json = serde_json::json!({
-            "reaction_type": {
-                "emoji_type": emoji_type
-            }
-        });
-        tracing::info!(
-            "[feishu:{}] add_reaction POST {} token={}... body={}",
-            bot_id,
-            url,
-            &token[..token.len().min(10)],
-            body_json
-        );
-        let res = match client
-            .post(&url)
-            .header("Authorization", format!("Bearer {token}"))
-            .json(&body_json)
-            .send()
-            .await
-        {
-            Ok(r) => r,
-            Err(e) => {
-                tracing::error!("[feishu:{}] add_reaction request failed: {e}", bot_id);
-                return None;
-            }
-        };
-
-        let status = res.status();
-        let body: serde_json::Value = match res.json().await {
-            Ok(b) => b,
-            Err(e) => {
-                tracing::error!("[feishu:{}] add_reaction parse failed: {e}", bot_id);
-                return None;
-            }
-        };
-
-        let code = body.get("code").and_then(|v| v.as_i64()).unwrap_or(-1);
-        if code != 0 {
-            tracing::error!(
-                "[feishu:{}] add_reaction API error (status={}): {body}",
-                bot_id,
-                status
-            );
-            return None;
-        }
-
-        let reaction_id = body
-            .get("data")
-            .and_then(|d| d.get("reaction_id"))
-            .and_then(|v| v.as_str())
-            .map(String::from);
-
-        tracing::info!(
-            "[feishu:{}] add_reaction {} ok, reaction_id={:?}",
-            bot_id,
-            emoji_type,
-            reaction_id
-        );
-        reaction_id
-    }
-
-    /// Delete reaction by reaction_id.
-    async fn delete_reaction(
-        credentials: &DashMap<i64, (String, String, String)>,
-        token_manager: &Arc<TokenManager>,
-        bot_id: i64,
-        message_id: &str,
-        reaction_id: &str,
-    ) {
-        let token = match Self::get_tenant_token(credentials, token_manager, bot_id).await {
-            Some(t) => t,
-            None => return,
-        };
-        let base_url = match Self::base_url(credentials, bot_id) {
-            Some(u) => u,
-            None => return,
-        };
-
-        let client = reqwest::Client::new();
-        match client
-            .delete(format!(
-                "{base_url}/open-apis/im/v1/messages/{message_id}/reactions/{reaction_id}"
-            ))
-            .header("Authorization", format!("Bearer {token}"))
-            .send()
-            .await
-        {
-            Ok(res) => {
-                let body: serde_json::Value = res.json().await.unwrap_or_default();
-                let code = body.get("code").and_then(|v| v.as_i64()).unwrap_or(-1);
-                if code == 0 {
-                    tracing::info!("[feishu:{}] delete_reaction ok", bot_id);
-                } else {
-                    tracing::error!("[feishu:{}] delete_reaction API error: {body}", bot_id);
-                }
-            }
-            Err(e) => {
-                tracing::error!("[feishu:{}] delete_reaction request failed: {e}", bot_id);
-            }
-        }
-    }
 }
 
 pub(crate) struct SlashCommandMatch<'a> {
@@ -2481,71 +910,9 @@ mod tests {
         assert_eq!(parsed.body, "");
     }
 
-    #[test]
-    fn test_parse_card_command_extracts_command_text() {
-        // 无参命令：cmd:/new → /new
-        assert_eq!(
-            FeishuListener::parse_card_command("cmd:/new"),
-            Some("/new".to_string())
-        );
-        // 带参命令：参数原样保留，交由后续 parse_slash_command 解析
-        assert_eq!(
-            FeishuListener::parse_card_command("cmd:/bind my-project"),
-            Some("/bind my-project".to_string())
-        );
-    }
 
-    #[test]
-    fn test_parse_card_command_returns_none_for_non_cmd_prefix() {
-        // nav:/act:/未知前缀/空串都不是命令点击，返回 None
-        assert_eq!(FeishuListener::parse_card_command("nav:/help common"), None);
-        assert_eq!(FeishuListener::parse_card_command("act:/delete-mode cancel"), None);
-        assert_eq!(FeishuListener::parse_card_command(""), None);
-    }
 
-    #[test]
-    fn test_parse_card_action_variants() {
-        use super::CardAction;
-        // 各 verb 正常解析（bind/runtodo/runloop 参数是 i64）
-        assert_eq!(FeishuListener::parse_card_action("act:/stop"), Some(CardAction::Stop));
-        assert_eq!(FeishuListener::parse_card_action("act:/new"), Some(CardAction::New));
-        assert_eq!(FeishuListener::parse_card_action("act:/bind 5"), Some(CardAction::Bind(5)));
-        assert_eq!(FeishuListener::parse_card_action("act:/runtodo 10"), Some(CardAction::RunTodo(10)));
-        assert_eq!(FeishuListener::parse_card_action("act:/runloop 20"), Some(CardAction::RunLoop(20)));
-        assert_eq!(
-            FeishuListener::parse_card_action("act:/push result_only"),
-            Some(CardAction::Push("result_only".to_string()))
-        );
-        // 缺参数 → None
-        assert_eq!(FeishuListener::parse_card_action("act:/bind"), None);
-        assert_eq!(FeishuListener::parse_card_action("act:/runtodo"), None);
-        assert_eq!(FeishuListener::parse_card_action("act:/push"), None);
-        // 非 i64 参数 / 未知 verb / 非 act 前缀 → None
-        assert_eq!(FeishuListener::parse_card_action("act:/bind abc"), None);
-        assert_eq!(FeishuListener::parse_card_action("act:/unknown"), None);
-        assert_eq!(FeishuListener::parse_card_action("nav:/help task"), None);
-        assert_eq!(FeishuListener::parse_card_action("cmd:/new"), None);
-    }
 
-    #[test]
-    fn test_resolve_receive_target_group_vs_private() {
-        use crate::feishu::ChannelMessage;
-        // 群聊：channel(chat_id)非空 → 用 chat_id 作为推送目标
-        let group_msg = ChannelMessage {
-            id: "om1".to_string(),
-            sender: "ou_user".to_string(),
-            sender_type: None,
-            content: "act:/stop".to_string(),
-            channel: "oc_group".to_string(),
-            timestamp: 0,
-            chat_type: Some("card_callback".to_string()),
-            mentioned_open_ids: vec![],
-        };
-        assert_eq!(FeishuListener::resolve_receive_target(&group_msg), ("oc_group", "chat_id"));
-        // 私聊：channel 空 → 回退到点击者 open_id
-        let private_msg = ChannelMessage { channel: String::new(), ..group_msg.clone() };
-        assert_eq!(FeishuListener::resolve_receive_target(&private_msg), ("ou_user", "open_id"));
-    }
 
     #[test]
     fn test_group_message_requires_mention_when_enabled() {
